Reset transitionAlpha when the transition is stopped early.
Bug 19273502

When a transition runs several animators and is interrupted,
some animators may not have been started. Those animators
that haven't been started will not reset their contents
the same as if the animator was canceled. Fade modifies
transitionAlpha and if it is interrupted prior to the
animator starting, the transitionAlpha doesn't get reset.

Change-Id: I8a49b4fbb7151aadc43d7b10e6acbae9578e5cc4