Start non-running activity that should be visible in resume state.
In multi-window mode it is possible for the system-ui launch
multiple activities at the same time that might not all be in a
running process. For the activities that are not in a running
process to be visible, they need to start in the resume state
not the stopped state if they will be the only resumed activity
in the stack.