Don't leak starting window when app crashes before adding windows
It is possible for us to leak an app starting window if the app it
is associated with fails to start before adding any windows. This
causes the starting window to permanently be visible at the layer
it occupies.

We now remove the starting window immediately when the app token
is been removed if it is the last window in the token and it isn't
animating.